fix parties
