+++ src/components/@parties/PartiesCard/index.jsx
@@ -1,31 +1,12 @@
-/* eslint-disable react/jsx-curly-spacing,no-restricted-globals */
 import React from 'react'
 import { object, func } from 'prop-types'
-import { withStyles } from '@material-ui/core/styles'
-import { Link } from 'react-router-dom'
-import FormControlLabel from '@material-ui/core/FormControlLabel/FormControlLabel'
-import Typography from '@material-ui/core/Typography/Typography'
-import Card from '@material-ui/core/Card'
-import CardHeader from '@material-ui/core/CardHeader'
-import CardContent from '@material-ui/core/CardContent'
-import CardActions from '@material-ui/core/CardActions'
-import Avatar from '@material-ui/core/Avatar'
-import IconButton from '@material-ui/core/IconButton'
-import Favorite from '@material-ui/icons/Favorite'
-import FavoriteBorder from '@material-ui/icons/FavoriteBorder'
-import ShareIcon from '@material-ui/icons/Share'
-import MoreVertIcon from '@material-ui/icons/MoreVert'
-import Button from '@material-ui/core/es/Button/Button'
-import PartiesCardDescription from './PartiesCardDescription'
-import Checkbox from '@material-ui/core/Checkbox/Checkbox'
-import Grid from '@material-ui/core/Grid/Grid'
-import TagList from './Tag/TagList'
-import moment from 'moment'
+import { withStyles, Card } from '@material-ui/core'
+import CardHeader from './Card/MyCardHeader'
+import CardContent from './Card/MyCardContent'
+import CardActions from './Card/MyCardActions'
 import connector from '../connector'
-import initialsFromUsername from 'src/utils/initialsFromUsername'
-import isEmpty from 'lodash/isEmpty'
 
-const styles = theme => ({
+const styles = {
   root: {
     marginBottom: 15,
     width: '100%',
@@ -33,49 +14,7 @@ const styles = theme => ({
       width: '95%',
     },
   },
-  avatarInitials: {
-    background: theme.palette.primary.main,
-  },
-  media: {
-    height: 0,
-    paddingTop: '56.25%', // 16:9
-  },
-  actions: {
-    display: 'flex',
-    marginTop: -30,
-  },
-  expand: {
-    transform: 'rotate(0deg)',
-    transition: theme.transitions.create('transform', {
-      duration: theme.transitions.duration.shortest,
-    }),
-    marginLeft: 'auto',
-  },
-  expandOpen: {
-    transform: 'rotate(180deg)',
-  },
-  flex: {
-    flexGrow: 1,
-    marginLeft: 30,
-  },
-  picture: {
-    width: '100%',
-    height: 200,
-    paddingBottom: 20,
-    borderRadius: 5,
-  },
-  title: {
-    paddingTop: 10,
-    marginLeft: '15%',
-    marginRight: '15%',
-  },
-  cardContent: {
-    paddingTop: 0,
-  },
-  type: {
-    display: 'flex',
-  },
-})
+}
 
 class PartiesCard extends React.Component {
   handleLikeClick = (id) => {
@@ -84,59 +23,11 @@ class PartiesCard extends React.Component {
 
   render() {
     const { classes, party } = this.props
-    const url = 'http://localhost:3333/images/parties.jpg'
     return (
       <Card className={classes.root}>
-        <CardHeader
-          avatar={party.admin.avatar_url
-            ? <Avatar src={party.admin.avatar_url} />
-            : <Avatar className={classes.avatarInitials}>{initialsFromUsername(party.admin.name)}</Avatar>
-          }
-          action={<IconButton><MoreVertIcon /></IconButton>}
-          title={party.admin.name}
-          subheader={moment(party.updated_at)
-            .fromNow()}
-        />
-        <CardContent className={classes.cardContent}>
-          <Link to={`/parties/${party.id}`}>
-            <Avatar
-              src={isEmpty(party.primary_picture) ? url : party.primary_picture}
-              onClick={() => this.handleClick(party)}
-              className={classes.picture}
-            />
-          </Link>
-          <div className={classes.type}>
-            <TagList iconTag={party.type} />
-            <Grid container justify="center" className={classes.title}>
-              <Link to={`/parties/${party.id}`}>
-                <Typography align="center" variant="title" color="primary">
-                  {party.title}
-                </Typography>
-              </Link>
-            </Grid>
-          </div>
-          <PartiesCardDescription party={party} />
-        </CardContent>
-        <CardActions className={classes.actions} disableActionSpacing>
-          <div className={classes.flex}>
-            <FormControlLabel
-              control={
-                <Checkbox
-                  icon={<FavoriteBorder />}
-                  checkedIcon={<Favorite color="primary" />}
-                  value="checked"
-                  onClick={() => this.handleLikeClick(party.id)}
-                />}
-              label="0"
-            />
-            <IconButton aria-label="Share">
-              <ShareIcon />
-            </IconButton>
-          </div>
-          <Link to={`/parties/${party.id}`}>
-            <Button color="primary">Подробнее</Button>
-          </Link>
-        </CardActions>
+        <CardHeader party={party} />
+        <CardContent party={party} />
+        <CardActions party={party} onLike={() => this.handleLikeClick(party.id)} />
       </Card>
     )
   }
